Header::NetBIOS: extract Session class to its own file.
